+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
+#### v1.3.0
+- Fixed an issue where hex-colored text in card descriptions ignored the alpha value of surrounding text
+- Modded tab bars in the Card Library now use a custom sprite instead of a cropped version of the regular long bar
+- Other minor edits
+
 #### v1.2.0
 - Added relocking functions
 - Added a few more unlock functions